+++ app/controllers/admin/users_controller.rb
@@ -1,9 +1,42 @@
 class Admin::UsersController < Admin::BaseController        
 
-  before_filter :find_user
+  before_filter :find_user, :except => :index
+  
   
   def index
-    @users = User.paginate :per_page => Tog::Config['plugins.tog_core.pagination_size'], :page => params[:page], :order => 'login'
+    @order_by = params[:order_by] || "login"
+    @sort_order = params[:sort_order] || "asc"
+    @page = params[:page] || '1'
+    
+    condition = ""
+    conditions_values = Hash.new
+    
+    if params[:search_term]
+      conditions_values[:search_term] = "%#{params[:search_term]}%"
+      condition = "(login like :search_term or email like :search_term)"
+    end
+    
+    if params[:age]
+      condition += " and " if condition != ""
+      today = Date.today
+      conditions_values[:aged] = {
+        :today => today,
+        :week => today - 7.days,
+        :month => today - 1.month,
+      }[params[:age].to_sym]
+      condition += "(created_at > :aged)"
+    end
+    
+    if params[:state]
+      condition += " and " if condition != ""
+      conditions_values[:state] = params[:state]
+      condition += "(state = :state)"
+    end    
+    
+    @users = User.find(:all, :order => "#{@order_by} #{@sort_order}",
+                             :conditions => [condition, conditions_values]
+                      ).paginate :page => @page,
+                                 :per_page => Tog::Config['plugins.tog_core.pagination_size']
   end
 
   def new