+++ test/test_helper.rb
@@ -20,4 +20,10 @@ begin require 'redgreen'; rescue LoadError; end
 class ActiveSupport::TestCase
   self.use_transactional_fixtures = true
   self.use_instantiated_fixtures  = false
+
+  # Check two arrays for set equality
+  def assert_set_equal(expected, actual)
+    assert_equal(expected.to_set, actual.to_set,
+                 "Expected: #{expected.inspect}, Actual: #{actual.inspect}")
+  end
 end