+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Edge
 ----
 
-0.5.3
-----
+* Filter users by creation date or state in admin (kudos to tokland)
+* Search users in admin (kudos to tokland)
 
 0.5.2
 ----